pyproject.toml
[project]
name = "Chinese_Grammar_Error_Correction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.1.2",
    "langchain-openai>=1.1.0",
    "pypdf>=6.4.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
